--- a/crates/rome_formatter/src/ts/auxiliary/initializer_clause.rs
+++ b/crates/rome_formatter/src/ts/auxiliary/initializer_clause.rs
@@ -1,9 +1,9 @@
 use crate::{
 	format_elements, space_token, FormatElement, FormatResult, Formatter, ToFormatElement,
 };
-use rslint_parser::ast::JsEqualValueClause;
+use rslint_parser::ast::JsInitializerClause;
 
-impl ToFormatElement for JsEqualValueClause {
+impl ToFormatElement for JsInitializerClause {
 	fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
 		Ok(format_elements![
 			formatter.format_token(&self.eq_token()?)?,