--- a/crates/rome_formatter/src/ts/auxiliary/mod.rs
+++ b/crates/rome_formatter/src/ts/auxiliary/mod.rs
@@ -1,5 +1,5 @@
-mod equal_value_clause;
 mod function_body;
+mod initializer_clause;
 mod name;
 mod reference_identifier;
 mod spread;