--- a/crates/rome_formatter/src/ts/statements/for_in_statement.rs
+++ b/crates/rome_formatter/src/ts/statements/for_in_statement.rs
@@ -1,43 +1,57 @@
-use rslint_parser::ast::{ForInStmt, ForLeft};
+use rslint_parser::ast::{JsAnyForInOrOfInitializer, JsForInStatement, JsForVariableDeclaration};
 
 use crate::{
 	format_elements, group_elements, soft_indent, soft_line_break_or_space, space_token,
 	FormatElement, FormatResult, Formatter, ToFormatElement,
 };
 
-impl ToFormatElement for ForInStmt {
+impl ToFormatElement for JsForInStatement {
 	fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
 		let for_token = formatter.format_token(&self.for_token()?)?;
 		let l_paren = formatter.format_token(&self.l_paren_token()?)?;
-		let left = formatter.format_node(self.left()?)?;
+		let initializer = formatter.format_node(self.initializer()?)?;
 		let in_token = formatter.format_token(&self.in_token()?)?;
-		let right = formatter.format_node(self.right()?)?;
+		let expression = formatter.format_node(self.expression()?)?;
 		let r_paren = formatter.format_token(&self.r_paren_token()?)?;
-		let cons = formatter.format_node(self.cons()?)?;
+		let body = formatter.format_node(self.body()?)?;
 
 		Ok(format_elements![
 			for_token,
 			space_token(),
 			l_paren,
 			group_elements(soft_indent(format_elements![
-				left,
+				initializer,
 				soft_line_break_or_space(),
 				in_token,
 				soft_line_break_or_space(),
-				right,
+				expression,
 			])),
 			r_paren,
 			space_token(),
-			cons
+			body
 		])
 	}
 }
 
-impl ToFormatElement for ForLeft {
+impl ToFormatElement for JsAnyForInOrOfInitializer {
 	fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
 		match self {
-			ForLeft::JsAnyAssignment(assignment) => assignment.to_format_element(formatter),
-			ForLeft::JsVariableDeclaration(decl) => decl.to_format_element(formatter),
+			JsAnyForInOrOfInitializer::JsAnyAssignmentPattern(assignment) => {
+				assignment.to_format_element(formatter)
+			}
+			JsAnyForInOrOfInitializer::JsForVariableDeclaration(decl) => {
+				decl.to_format_element(formatter)
+			}
 		}
+	}
+}
+
+impl ToFormatElement for JsForVariableDeclaration {
+	fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
+		Ok(format_elements![
+			formatter.format_token(&self.kind_token()?)?,
+			space_token(),
+			formatter.format_node(self.declaration()?)?
+		])
 	}
 }