--- a/crates/rome_formatter/src/ts/statements/for_of_statement.rs
+++ b/crates/rome_formatter/src/ts/statements/for_of_statement.rs
@@ -1,34 +1,34 @@
-use rslint_parser::ast::ForOfStmt;
+use rslint_parser::ast::JsForOfStatement;
 
 use crate::{
 	format_elements, group_elements, soft_indent, soft_line_break_or_space, space_token,
 	FormatElement, FormatResult, Formatter, ToFormatElement,
 };
 
-impl ToFormatElement for ForOfStmt {
+impl ToFormatElement for JsForOfStatement {
 	fn to_format_element(&self, formatter: &Formatter) -> FormatResult<FormatElement> {
 		let for_token = formatter.format_token(&self.for_token()?)?;
 		let l_paren = formatter.format_token(&self.l_paren_token()?)?;
-		let left = formatter.format_node(self.left()?)?;
+		let initializer = formatter.format_node(self.initializer()?)?;
 		let of_token = formatter.format_token(&self.of_token()?)?;
-		let right = formatter.format_node(self.right()?)?;
+		let expression = formatter.format_node(self.expression()?)?;
 		let r_paren = formatter.format_token(&self.r_paren_token()?)?;
-		let cons = formatter.format_node(self.cons()?)?;
+		let body = formatter.format_node(self.body()?)?;
 
 		Ok(format_elements![
 			for_token,
 			space_token(),
 			l_paren,
 			group_elements(soft_indent(format_elements![
-				left,
+				initializer,
 				soft_line_break_or_space(),
 				of_token,
 				soft_line_break_or_space(),
-				right,
+				expression,
 			])),
 			r_paren,
 			space_token(),
-			cons
+			body
 		])
 	}
 }