--- a/crates/rslint_parser/src/syntax/class.rs
+++ b/crates/rslint_parser/src/syntax/class.rs
@@ -719,7 +719,7 @@ fn parse_property_class_member_body(p: &mut Parser, member_marker: Marker) -> Co
 	}
 
 	maybe_ts_type_annotation(p);
-	parse_equal_value_clause(p).or_missing(p);
+	parse_initializer_clause(p).or_missing(p);
 
 	if !optional_semi(p) {
 		// Gets the start of the member
@@ -759,14 +759,14 @@ fn optional_member_token(p: &mut Parser) -> Option<Range<usize>> {
 	}
 }
 
-pub(crate) fn parse_equal_value_clause(p: &mut Parser) -> ParsedSyntax<CompletedMarker> {
+pub(crate) fn parse_initializer_clause(p: &mut Parser) -> ParsedSyntax<CompletedMarker> {
 	if p.at(T![=]) {
 		let m = p.start();
 		p.bump(T![=]);
 
 		expr_or_assignment(p);
 
-		Present(m.complete(p, JS_EQUAL_VALUE_CLAUSE))
+		Present(m.complete(p, JS_INITIALIZER_CLAUSE))
 	} else {
 		Absent
 	}