--- a/crates/rslint_parser/test_data/inline/err/for_stmt_err.js
+++ b/crates/rslint_parser/test_data/inline/err/for_stmt_err.js
@@ -5,3 +5,4 @@ for (in []) {}
 for (let i, j = 6 of []) {}
 for await (let a in []) {}
 for await (let i = 0; i < 10; ++i) {}
+for (let [a];;) {}