… crates/rslint_parser/src/ast/generated/nodes.rs
… crates/rslint_parser/src/syntax/util.rs
… crates/rslint_parser/src/util.rs
… crates/rslint_parser/test_data/inline/err/array_binding_err.rast
… crates/rslint_parser/test_data/inline/err/array_binding_rest_err.rast
… crates/rslint_parser/test_data/inline/err/async_arrow_expr_await_parameter.rast
… crates/rslint_parser/test_data/inline/err/binding_identifier_invalid.rast
… crates/rslint_parser/test_data/inline/err/binding_identifier_invalid_script.rast
… crates/rslint_parser/test_data/inline/err/class_declare_member.rast
… crates/rslint_parser/test_data/inline/err/debugger_stmt.rast
… crates/rslint_parser/test_data/inline/err/for_stmt_err.rast
… crates/rslint_parser/test_data/inline/err/import_call_no_arg.rast
… crates/rslint_parser/test_data/inline/err/invalid_arg_list.rast
… crates/rslint_parser/test_data/inline/err/invalid_method_recover.rast
… crates/rslint_parser/test_data/inline/err/object_binding_pattern.rast
… crates/rslint_parser/test_data/inline/err/object_expr_err.rast
… crates/rslint_parser/test_data/inline/err/object_expr_error_prop_name.rast
… crates/rslint_parser/test_data/inline/err/object_expr_method.rast
… crates/rslint_parser/test_data/inline/err/object_expr_non_ident_literal_prop.rast
… crates/rslint_parser/test_data/inline/err/object_expr_setter.rast
… crates/rslint_parser/test_data/inline/err/object_property_binding_err.rast
… crates/rslint_parser/test_data/inline/err/object_shorthand_property_err.rast
… crates/rslint_parser/test_data/inline/err/primary_expr_invalid_recovery.rast
… crates/rslint_parser/test_data/inline/err/property_assignment_target_err.rast
… crates/rslint_parser/test_data/inline/err/rest_property_binding_err.rast
… crates/rslint_parser/test_data/inline/err/semicolons_err.rast
… crates/rslint_parser/test_data/inline/err/template_literal_unterminated.rast
… crates/rslint_parser/test_data/inline/err/unterminated_unicode_codepoint.rast
… crates/rslint_parser/test_data/inline/err/var_decl_err.rast
… crates/rslint_parser/test_data/inline/err/variable_declaration_statement_err.rast