## Summary

- Renames `EqualValueClause` to `JsInitializerClause`
- Renames `JsVariableDeclarationStatement` to `JsVariableStatement`
- Renames `JsVariableDeclaration` to `JsVariableDeclarationList`
- Rename `JsVariableDeclarator` to `JsVariableDeclaration`
- Rename `ForInStmt` to `JsForInStatement`
- Rename `ForOfStmt` to `JsForOfStatement`
- Delete `ForStmtInit`
- Introduce `JsForVariableDeclaration` that only allows a single variable declaration.

Deletes the `parse_marker` from the `Parser`.

This PR doesn't change the error recovery for `for` statements nor for variable declarations because of the pending discussion in #1867. I'll spend some time now to prototype an approach that uses a generic `Unknown` to remove conditional syntax.

## Test Plan

The parser's performance seems to improve a bit. I suspect this is stemming from the removal of the `parse_marker` to verify if the variable declaration of a for of/in loop only contains a single variable.


* jquery: 6.98ms -> 6.49ms
* vue: 13.5788ms -> 12.194ms
* react: 496.9µs -> 461.5µs
* react-dom: 8.6984ms -> 7.7618ms
* compiler: 26.9218ms -> 22.8185ms
* dojo:  1.8681ms -> 1.6019ms
* d3: 23.1232ms -> 19.6941ms
* pixi: 27.6754ms -> 23.5175ms
* three: 31.8306ms -> 28.224ms
* tex: 73.1209ms -> 46.4471ms